1. Заменена обработки всех исключительных ситуаций на выброс исклюючение RuntimeException

+++ src/main/java/service/TaskService.java
@@ -14,40 +14,35 @@ public class TaskService {
     }
 
     //CREATE
-    public void create(String id, String name, String description, String dateStart, String dateFinish, String projectId){
+    public Task create(String id, String name, String description, String dateStart, String dateFinish, String projectId){
         int idChecked;
         if(id == null) idChecked = taskRepository.getRepositorySize();
         else idChecked = Integer.parseInt(id);
 
         if(name == null || name.isEmpty()){
-            System.out.println("Имя проекта не может быть пустым");
-            return;
+            throw new RuntimeException("Имя задачи не может быть пустым");
         }
         if(dateStart == null || dateStart.isEmpty()){
-            System.out.println("Дата начала не может быть пустой");
-            return;
+            throw new RuntimeException("Дата начала задачи не может быть пустой");
         }
         if(dateFinish == null || dateFinish.isEmpty()){
-            System.out.println("Дата окончания не может быть пустой");
-            return;
+            throw new RuntimeException("Дата окончания задачи не может быть пустой");
         }
         int projectIdChecked;
         if(projectId == null || projectId.isEmpty()){
-            System.out.println("Пользователь проекта не может быть пустой");
-            return;
+            throw new RuntimeException("Id проекта не может быть пустым");
         }
         else{
             projectIdChecked = Integer.parseInt(projectId);
         }
 
-        taskRepository.create(idChecked, name, description, LocalDate.parse(dateStart), LocalDate.parse(dateFinish), projectIdChecked);
+        return taskRepository.create(idChecked, name, description, LocalDate.parse(dateStart), LocalDate.parse(dateFinish), projectIdChecked);
     }
 
     //READ
     public Task findByName(String taskName){
         if(taskName == null || taskName.isEmpty()){
-            System.out.println("Имя задачи не может быть пустым");
-            return null;
+            throw new RuntimeException("Имя задачи не может быть пустым");
         }
         return taskRepository.findByName(taskName);
     }
@@ -58,41 +53,38 @@ public class TaskService {
     //UPDATE
     public void updateByName(String taskName, String fieldForUpdate, String newValue){
         if(taskName == null || taskName.isEmpty()){
-            System.out.println("Имя задачи не может быть пустым");
-            return;
+            throw new RuntimeException("Имя задачи не может быть пустым");
         }
         if(fieldForUpdate == null || fieldForUpdate.isEmpty()){
-            System.out.println("Поле для обновления не может быть пустой");
-            return;
+            throw new RuntimeException("Поле для обновления не может быть пустым");
         }
         if((newValue == null || newValue.isEmpty()) && !fieldForUpdate.equalsIgnoreCase("описание")){
-            System.out.println("Новое значение не может быть пустой");
-            return;
+            throw new RuntimeException("Новое значение не может быть пустым");
         }
         taskRepository.updateByName(taskName, fieldForUpdate, newValue);
     }
     public void updateById(String taskId, String fieldForUpdate, String newValue){
         int taskIdChecked;
         if(taskId == null || taskId.isEmpty()){
-            System.out.println("ID проекта не может быть пустым");
-            return;
+            throw new RuntimeException("ID проекта не может быть пустым");
         }
         else{
             taskIdChecked = Integer.parseInt(taskId);
         }
         if(fieldForUpdate == null || fieldForUpdate.isEmpty()){
-            System.out.println("Поле для обновления не может быть пустой");
-            return;
+            throw new RuntimeException("Поле для обновления не может быть пустым");
         }
         if((newValue == null || newValue.isEmpty()) && !fieldForUpdate.equalsIgnoreCase("описание")){
-            System.out.println("Новое значение не может быть пустой");
-            return;
+            throw new RuntimeException("Новое значение не может быть пустым");
         }
         taskRepository.updateById(taskIdChecked, fieldForUpdate, newValue);
     }
 
     //DELETE
     public void deleteByName(String taskName){
+        if(taskName == null || taskName.isEmpty()){
+            throw new RuntimeException("Имя задачи не может быть пустым");
+        }
         taskRepository.deleteByName(taskName);
     }
 }
